create shop py to store
