+++ app/__init__.py
@@ -20,8 +20,10 @@ def create_app():
 
     from app.routes import bp
     from app.inventory import bp as inventory_bp
+    from app.shop import bp as shop_bp
     app.register_blueprint(bp)
     app.register_blueprint(inventory_bp)
+    app.register_blueprint(shop_bp)
     register_cli(app)
 
     return app
