+++ app/templates/base.html
@@ -37,7 +37,7 @@
                         </li>
                         <li class="nav-item">
                             <a class="nav-link {% if request.endpoint == 'shop' %}active{% endif %}"
-                                href="{{ url_for('main.shop') }}">Shop</a>
+                                href="{{ url_for('shop.shop') }}">Shop</a>
                         </li>
                         <li class="nav-item">
                             <span class="nav-link">
